Add support for destination volumes that are not zerored
This change adds the destination_assume_zeroed variable which
will remove the --destination-is-zero flag from nbdcopy calls. The
default behavior if using the binary migrate tool directly has changed
unless assumezero is included in the input JSON data.

When Cinder volume encryption is enabled for the destination volume
type, the target volume will be populated with random data instead of
being zerored. This causes disk corruption when nbdkit migrates data
with the --destination-is-zero flag.